Реализована функция обработки данных и исправлены ошибки

+++ main-project/process.cpp
@@ -2,7 +2,7 @@
 int periodToSeconds(const phone_subscription* p) {
     return p->tm.hours * 3600 + p->tm.minutes * 60 + p->tm.seconds;
 }
-int process(phone_subscription* array[], int size)
+double process(phone_subscription* array[], int size)
 {
     double totalCost = 0.0;
     double totalSeconds = 0.0;
